refactor(console): improve tip component usage
