+++ packages/console/src/components/FormField/index.module.scss
@@ -17,7 +17,7 @@
   }
 
   .toggleTipButton {
-    margin-left: _.unit(1);
+    margin-left: _.unit(0.5);
   }
 
   .required {
