+++ packages/console/src/components/IconButton/index.tsx
@@ -1,18 +1,16 @@
 import classNames from 'classnames';
-import type { ForwardedRef, HTMLProps, ReactNode } from 'react';
+import type { ForwardedRef, HTMLProps } from 'react';
 import { forwardRef, useRef } from 'react';
 
-import Tooltip from '../Tooltip';
 import * as styles from './index.module.scss';
 
 export type Props = Omit<HTMLProps<HTMLButtonElement>, 'size' | 'type'> & {
   size?: 'small' | 'medium' | 'large';
-  tooltip?: ReactNode;
   iconClassName?: string;
 };
 
 const IconButton = (
-  { size = 'medium', children, className, tooltip, iconClassName, ...rest }: Props,
+  { size = 'medium', children, className, iconClassName, ...rest }: Props,
   reference: ForwardedRef<HTMLButtonElement>
 ) => {
   const tipRef = useRef<HTMLDivElement>(null);
@@ -27,9 +25,6 @@ const IconButton = (
       <div ref={tipRef} className={classNames(styles.icon, iconClassName)}>
         {children}
       </div>
-      {tooltip && (
-        <Tooltip anchorRef={tipRef} content={tooltip} position="top" horizontalAlign="center" />
-      )}
     </button>
   );
 };
