+++ packages/console/src/components/Tip/TipBubble/index.module.scss
