+++ packages/console/src/components/Tip/Tooltip/index.module.scss
