+++ packages/console/src/components/Tip/index.ts
@@ -1,0 +1,2 @@
+export { default as Tooltip } from './Tooltip';
+export { default as ToggleTip } from './ToggleTip';
