+++ packages/console/src/pages/ConnectorDetails/components/SenderTester/index.tsx
@@ -1,7 +1,8 @@
 import { phoneRegEx, emailRegEx } from '@logto/core-kit';
 import { ConnectorType } from '@logto/schemas';
+import { conditional } from '@silverhand/essentials';
 import classNames from 'classnames';
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { toast } from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
@@ -9,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 import Button from '@/components/Button';
 import FormField from '@/components/FormField';
 import TextInput from '@/components/TextInput';
-import Tooltip from '@/components/Tooltip';
+import { Tooltip } from '@/components/Tip';
 import useApi from '@/hooks/use-api';
 import { safeParseJson } from '@/utilities/json';
 
@@ -27,7 +28,6 @@ type FormData = {
 };
 
 const SenderTester = ({ connectorId, connectorType, config, className }: Props) => {
-  const buttonPosReference = useRef(null);
   const [showTooltip, setShowTooltip] = useState(false);
   const {
     handleSubmit,
@@ -100,23 +100,19 @@ const SenderTester = ({ connectorId, connectorType, config, className }: Props) 
             })}
           />
         </FormField>
-        <div ref={buttonPosReference} className={styles.send}>
+        <Tooltip
+          isKeepOpen
+          anchorClassName={styles.send}
+          className={styles.successfulTooltip}
+          content={conditional(showTooltip && t('connector_details.test_message_sent'))}
+        >
           <Button
             isLoading={isSubmitting}
             title="connector_details.send"
             type="outline"
             onClick={onSubmit}
           />
-        </div>
-        {showTooltip && (
-          <Tooltip
-            isKeepOpen
-            horizontalAlign="center"
-            className={styles.successfulTooltip}
-            anchorRef={buttonPosReference}
-            content={t('connector_details.test_message_sent')}
-          />
-        )}
+        </Tooltip>
       </div>
       <div className={classNames(inputError?.message ? styles.error : styles.description)}>
         {inputError?.message ?? t('connector_details.test_sender_description')}
