+++ packages/console/src/pages/Connectors/components/ConnectorStatusField/index.module.scss
@@ -5,7 +5,7 @@
   align-items: center;
 
   .tipButton {
-    margin-left: _.unit(1);
+    margin-left: _.unit(0.5);
   }
 }
