+++ packages/console/src/pages/Dashboard/components/Block.module.scss
@@ -35,7 +35,7 @@
     align-items: center;
 
     .toggleTipButton {
-      margin-left: _.unit(1);
+      margin-left: _.unit(0.5);
     }
   }
