+++ packages/console/src/pages/Dashboard/components/Block.tsx
@@ -5,8 +5,10 @@ import { useTranslation } from 'react-i18next';
 
 import ArrowDown from '@/assets/images/arrow-down.svg';
 import ArrowUp from '@/assets/images/arrow-up.svg';
+import Tip from '@/assets/images/tip.svg';
 import Card from '@/components/Card';
-import ToggleTipButton from '@/components/ToggleTipButton';
+import IconButton from '@/components/IconButton';
+import { ToggleTip } from '@/components/Tip';
 import { formatNumberWithComma } from '@/utilities/number';
 
 import * as styles from './Block.module.scss';
@@ -29,7 +31,11 @@ const Block = ({ variant = 'default', count, delta, title, tip }: Props) => {
       <div className={styles.title}>
         {t(title)}
         {tip && (
-          <ToggleTipButton className={styles.toggleTipButton} render={() => <div>{t(tip)}</div>} />
+          <ToggleTip anchorClassName={styles.toggleTipButton} content={<div>{t(tip)}</div>}>
+            <IconButton size="small">
+              <Tip />
+            </IconButton>
+          </ToggleTip>
         )}
       </div>
       <div className={styles.content}>
